Refactor naming, add comments
